requirements.txt
moviepy==1.0.3
fuzzysearch==0.8.1
openai-whisper
